fix bad imports from src/core/http.ts
